splitted types
