+++ src/internal.ts
@@ -1,6 +1,7 @@
 export * from "./env"
 export * from "./extends"
-export * from "./types"
+export * from "./types-external"
+export * from "./types-internal"
 export * from "./common"
 export * from "./scope"
 export * from "./finalize"
